+++ server/src/covet/api/__init__.py
@@ -7,6 +7,7 @@ from covet.api import auth as auth_router
 from covet.api import categories as categories_router
 from covet.api import chores as chores_router
 from covet.api import collections as collections_router
+from covet.api import comments as comments_router
 from covet.api import contacts as contacts_router
 from covet.api import documents as documents_router
 from covet.api import ha as ha_router
@@ -35,6 +36,7 @@ api_router.include_router(auth_router.router)
 api_router.include_router(collections_router.router)
 api_router.include_router(categories_router.router)
 api_router.include_router(items_router.router)
+api_router.include_router(comments_router.router)
 api_router.include_router(inventory_router.router)
 api_router.include_router(item_templates_router.router)
 api_router.include_router(tags_router.router)