+++ server/src/covet/models/__init__.py
@@ -12,6 +12,7 @@ from covet.models.contact import Contact
 from covet.models.document import Document
 from covet.models.invitation import Invitation
 from covet.models.item import Item
+from covet.models.item_comment import ItemComment
 from covet.models.item_lot import ItemLot
 from covet.models.item_template import ItemTemplate
 from covet.models.loan import Loan
@@ -57,6 +58,7 @@ __all__ = [
     "Document",
     "Invitation",
     "Item",
+    "ItemComment",
     "ItemLot",
     "ItemTag",
     "ItemTemplate",